teste imagensPath

+++ CadastroClientes/Store/ImagensChecklist.cs
@@ -19,6 +19,10 @@
             var safeName = $"{DateTime.UtcNow:yyyyMMdd_HHmmssfff}_{Guid.NewGuid():N}{Path.GetExtension(nomeArquivo)}";
 
             var caminhoCompleto = Path.Combine(imagensPath, safeName);
+
+            if (!File.Exists(imagensPath))
+                throw new Exception($"Imagem não encontrada: {imagensPath}");
+
             File.WriteAllBytes(caminhoCompleto, imagemBytes);
 
             return caminhoCompleto; // caminho absoluto!
